Fixed bug Reporting-15. Image table report does not display Background image
I also added the code to support the BackgroundRepeat report property. All the values defined in the "Report Page Layout (RPL) ..." are supported and translated into HTML styles per the BackgroundRepeatTypesMap map.

This main fix was in the method GetNonTextStyle. But I had to refactor a number of the function signatures so that I had the ReportState when I needed it. I also changed the WriteImage function as the same code snippet was used there (I.e., the REST call to GetImage).

Signed Off By: Jon Totton <jont@forerunnersw.com>